fix: resolve Linux build errors and various compiler warnings
- Fixed build error in texture.cpp by wrapping GetLastError() in WIN32 guards.
- Initialized 'obstacle' in physics.cpp to fix -Wmaybe-uninitialized.
- Added bounds checks to memset calls in crypto.cpp and worldio.cpp to resolve -Wstringop-overflow.
- Updated clipconsole signature in game.h to match derived classes and fix -Woverloaded-virtual.
- Refactored macros and used size_t for counts in obj.h and animmodel.h to address stringop-overflow warnings.
- Updated octarender.cpp to use explicit size_t length for memcpy.

+++ src/game/game.h
@@ -685,7 +685,7 @@ namespace game
         virtual ~clientmode() {}
 
         virtual void preload() {}
-        virtual float clipconsole(int w, int h) { return 0; }
+        virtual float clipconsole(float w, float h) { return 0; }
         virtual void drawhud(fpsent *d, int w, int h) {}
         virtual void rendergame() {}
         virtual void respawned(fpsent *d) {}
